requirements.txt
pandas
sqlalchemy
psycopg2-binary
python-dotenv
apache-airflow
pytest
jupyter
pandas
numpy
apache-airflow-providers-postgres
